requirements.txt
asgiref==3.7.2
Django==4.2.2
gunicorn==20.1.0
psycopg2==2.9.6
sqlparse==0.4.4
whitenoise==6.4.0
djangorestframework==3.14.0
pdf2image~=1.16.3
locust~=2.15.1